Fixing bug for training
