+++ yolort/v5/__init__.py
@@ -1,3 +1,5 @@
 from .models import *
 from .utils import *
 from .helper import *
+
+from .train import run
